### 2021-05-05 #3248 @whedon
[PRE REVIEW]: OpenOmics: A bioinformatics API to integrate multi-omics datasets and interface with public databases
**Submitting author:** @JonnyTran (<a href="http://orcid.org/0000-0002-2575-9633">Nhat Tran</a>)
**Repository:** <a href="https://github.com/BioMeCIS-Lab/OpenOmics" target ="_blank">https://github.com/BioMeCIS-Lab/OpenOmics</a>
**Version:** v0.8.8
