### 2021-05-05 #3249 @arfon
Fast-tracking this submission because it has already been reviewed in https://github.com/pyOpenSci/software-review/issues/31#issuecomment-832451199
