### 2021-05-08 #3249 @arfon
@JonnyTran - please merge this PR which makes a number of small formatting fixes to your paper: https://github.com/BioMeCIS-Lab/OpenOmics/pull/164

Also, please be sure to cite Dask and Pandas in your paper. You can find information on how to cite them here:

- https://docs.dask.org/en/latest/cite.html
- https://pandas.pydata.org/about/citing.html
