### 2021-05-08 #3249 @JonnyTran
Hi Arfon, I've just added pandas and dask citations.

I opted for the bibtex with author names,
> @InProceedings{ matthew_rocklin-proc-scipy-2015,
  author    = { Matthew Rocklin },
  title     = { Dask: Parallel Computation with Blocked algorithms and Task Scheduling },
  booktitle = { Proceedings of the 14th Python in Science Conference },
  pages     = { 130 - 136 },
  year      = { 2015 },
  editor    = { Kathryn Huff and James Bergstra }
}

rather than team authors,
> @Manual{,
  title = {Dask: Library for dynamic task scheduling},
  author = {{Dask Development Team}},
  year = {2016},
  url = {https://dask.org},
}

because Whedon will pick up the last name as (Team, 2016) citation.

Thanks for checking this over. I hope I didn't miss anything else.
